refactor: rename state-related 'expandedModal' to 'QuestionnaireExpandedModal'
