+++ src/components/common/Questionnaire/CollapsedDrawer/index.js
@@ -1,13 +1,13 @@
 import React, { useCallback } from 'react';
 import styles from './CollapsedDrawer.module.css';
 import { LS_USER_FEEDBACK_SUBMISSION_TIME_KEY } from 'constants/localStorageKey';
-import { toggleModalOpen } from 'actions/expandedModal';
+import { toggleModalOpen } from 'actions/questionnaireExpandedModal';
 import { useDispatch, useSelector } from 'react-redux';
 
 const MILLISECONDS_IN_30_DAYS = 1000 * 60 * 60 * 24 * 30;
 
 const CollapsedDrawer = ({ title = '給我們回饋', children }) => {
-  const isOpen = useSelector(state => state.expandedModal.isOpen);
+  const isOpen = useSelector(state => state.questionnaireExpandedModal.isOpen);
   const dispatch = useDispatch();
   const handleToggleModalOpen = useCallback(() => {
     dispatch(toggleModalOpen());
