--- src/selectors/expandedModal.js
@@ -1,3 +1,0 @@
-import { path } from 'ramda';
-
-export const expandedModal = path(['expandedModal']);
