+++ src/selectors/questionnaireExpandedModal.js
@@ -1,0 +1,3 @@
+import { path } from 'ramda';
+
+export const questionnaireExpandedModal = path(['questionnaireExpandedModal']);
